+++ packages/fighting-design/badge/src/interface.d.ts
@@ -1,4 +1,1 @@
 export type { BadgeProps } from './props'
-
-/** 不同类型 */
-export type BadgeType = 'primary' | 'success' | 'danger' | 'warning'