+++ start/src/App.vue
@@ -1,17 +1,21 @@
 <script lang="ts" setup></script>
 
 <template>
-  <f-alert round type="default">默认提示信息</f-alert>
-  <f-alert type="primary">主要提示信息</f-alert>
-  <f-alert type="success">成功提示信息</f-alert>
-  <f-alert type="danger">危险提示信息</f-alert>
-  <f-alert type="warning">警告提示信息</f-alert>
-
-  <f-alert :font-size="12" type="default">默认提示信息</f-alert>
-  <f-alert :font-size="14" type="primary">主要提示信息</f-alert>
-  <f-alert :font-size="16" type="success">成功提示信息</f-alert>
-  <f-alert font-size="20px" type="danger">危险提示信息</f-alert>
-  <f-alert font-size="24px" type="warning">警告提示信息</f-alert>
+  <f-badge value="11" type="default">
+    <f-button>default</f-button>
+  </f-badge>
+  <f-badge value="12" type="primary">
+    <f-button>primary</f-button>
+  </f-badge>
+  <f-badge value="13" type="success">
+    <f-button>success</f-button>
+  </f-badge>
+  <f-badge value="14" type="warning">
+    <f-button>warning</f-button>
+  </f-badge>
+  <f-badge value="14" type="danger">
+    <f-button>warning</f-button>
+  </f-badge>
 </template>
 
 <style lang="scss" scoped></style>