+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+- `f-badge` 组件 type 新增 default 配置项
+
 ## 0.20.3 (2023-01-31)
 
 - 针对 `f-button` 组件样式做了较大优化，内部样式实现优化